Add files via upload
This Power BI Dashboard for Hotel Reservations utilizes a dataset that contains information on both Resort and City hotels from the years 2015 to 2017. 
The analysis focuses on several key aspects:

1. Extracting insights specifically from the year 2015.
2. Concentrating on hotels of the Resort type.
3. Examining cancellations on a monthly basis.
4. Calculating the average daily rate.
5. Comparing the monthly revenue generated to the monthly revenue loss.
6. To enhance performance, a look-up table is employed in the report. 
7. The report provides information on the total revenue generated, total revenue lost, and the total number of cancellations.

Furthermore, the report allows users to analyze the data by breaking it down into months, quarters, and days, offering flexibility in exploring the dataset.